M    TERsrc/stringInfosHasher.cpp M    TERsrc/stringHasher.cpp M    TERsrc/stringInfosHasher.h M    TERsrc/stringHasher.h M    TERsrc/hashMapStringInfos.cpp M    TERsrc/infosHasher.cpp M    TERsrc/hashMapStringInfos.h M    TERsrc/infosHasher.h M    TERsrc/hashMapInfos.cpp M    TERsrc/hashMap.cpp M    TERsrc/hashMapInfos.h M    TERsrc/hashMap.h
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/branches/mert-other_metrics@3815 1f5c12ca-751b-0410-a591-d2e778427230

+++ mert/TERsrc/hashMapStringInfos.h
@@ -21,8 +21,8 @@ namespace HashMapSpace
 
         public:
 //     ~hashMap();
-            size_t hashValue ( string key );
-            int trouve ( size_t searchKey );
+            long hashValue ( string key );
+            int trouve ( long searchKey );
             int trouve ( string key );
             void addHasher ( string key, vector<string>  value );
             void addValue ( string key, vector<string>  value );
